dom/media/mediasource/test/test_SetModeThrows.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: append initialization</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

// MSE supports setting mode now. make sure it does not throw.
runWithMSE(function(ms, v) {
  ms.addEventListener("sourceopen", () => {
    const sb = ms.addSourceBuffer("video/webm");

    sb.mode = "segments";
    ok("true", "Setting to segments does not throw");
    try {
      sb.mode = "sequence";
      ok("true", "Setting to sequence does not throw");
    } catch (e) { ok(false, "Should not throw setting mode to sequence: " + e); }

    SimpleTest.finish();
  });
});

</script>
</pre>
</body>
</html>
